Rename crypto to ed25519 and use key at the module name.
