+++ lib/Ed25519VerificationKey2020.js
@@ -2,7 +2,7 @@
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
 import * as base58btc from 'base58-universal';
-import crypto from './crypto';
+import key from './ed25519';
 import {LDVerifierKeyPair} from 'crypto-ld';
 import {randomBytes} from '@stablelib/random';
 
@@ -74,9 +74,9 @@ export class Ed25519VerificationKey2020 extends LDVerifierKeyPair {
       secureRandom = () => randomBytes(32);
     }
     if(seed) {
-      keyObject = await crypto.generateKeyPairFromSeed(seed);
+      keyObject = await key.generateKeyPairFromSeed(seed);
     } else {
-      keyObject = await crypto.generateKeyPair({
+      keyObject = await key.generateKeyPair({
         randomBytes: secureRandom
       });
     }
@@ -227,7 +227,7 @@ export class Ed25519VerificationKey2020 extends LDVerifierKeyPair {
 
     return {
       async sign({data}) {
-        return crypto.sign(privateKeyBuffer, data);
+        return key.sign(privateKeyBuffer, data);
       },
     };
   }
@@ -237,7 +237,7 @@ export class Ed25519VerificationKey2020 extends LDVerifierKeyPair {
 
     return {
       async verify({data, signature}) {
-        return crypto.verify(publicKeyBuffer, data, signature);
+        return key.verify(publicKeyBuffer, data, signature);
       },
     };
   }
