+++ lib/ed25519-browser.js
