+++ lib/ed25519.js
